rtc-syntacore: add helper functions for setting time speed
I've noticed that in proc_write_spd() in randome mode time_stamp and
start_time updates 2 times in a row. It is not a big deal, but still...
So, this commit introduces 2 separate functions for setting time speed
coefficient (also updates random time speed), and for setting random
time speed coefficient alone.